relocate comments template to film detail
