+++ src/components/film-detail.js
@@ -1,8 +1,31 @@
 import AbstractSmartComponent from "./abstract-smart-component.js";
+import {generateCountObjects, generateComment} from "../mock/film.js";
 
-const filmDetailTemplate = (film, count, details) => {
+const createCommentsTemplate = (author, text, emoji, dayCommented) => {
+  return (
+    `<li class="film-details__comment">
+    <span class="film-details__comment-emoji">
+      <img src="./images/emoji/${emoji}" width="55" height="55" alt="emoji-smile">
+    </span>
+    <div>
+      <p class="film-details__comment-text"> ${text}</p>
+      <p class="film-details__comment-info">
+        <span class="film-details__comment-author">${author}</span>
+        <span class="film-details__comment-day">${dayCommented}</span>
+        <button class="film-details__comment-delete">Delete</button>
+      </p>
+    </div>
+  </li>`
+  );
+};
+
+
+const filmDetailTemplate = (film, count, details, emoji) => {
   const {title, rating, filmPublicationDate, duration, genre, img, description, monthPublicationDate, datePublication, isWatchList, isWatched, isFavorite} = film;
   const {actors, director, writers, country, filmDetailsAge} = details;
+  const emojiMarkup = emoji ? `<img src ="${emoji}" alt="" width="55" height="55">` : ``;
+  const comments = generateCountObjects(count, generateComment);
+  const commentsMarkup = comments.map((comment) => createCommentsTemplate(comment.author, comment.text, comment.emoji, comment.dayCommented)).join(`\n`);
   return (
     `<section class="film-details">
     <form class="film-details__inner" action="" method="get">
@@ -84,6 +107,39 @@ const filmDetailTemplate = (film, count, details) => {
         <section class="film-details__comments-wrap">
         <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${count}</span></h3>
         <ul class="film-details__comments-list">
+          ${commentsMarkup}
+
+          <div class="film-details__new-comment">
+    <div for="add-emoji" class="film-details__add-emoji-label">
+    ${emojiMarkup}
+    <input type="hidden" name="add-emoji" value="">
+    </div>
+
+    <label class="film-details__comment-label">
+      <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment"></textarea>
+    </label>
+
+    <div class="film-details__emoji-list">
+      <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-smile" value="smile">
+      <label class="film-details__emoji-label" for="emoji-smile">
+        <img src="./images/emoji/smile.png" width="30" height="30" alt="emoji">
+      </label>
+
+      <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-sleeping" value="sleeping">
+      <label class="film-details__emoji-label" for="emoji-sleeping">
+        <img src="./images/emoji/sleeping.png" width="30" height="30" alt="emoji">
+      </label>
+
+      <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-puke" value="puke">
+      <label class="film-details__emoji-label" for="emoji-puke">
+        <img src="./images/emoji/puke.png" width="30" height="30" alt="emoji">
+      </label>
+
+      <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-angry" value="angry">
+      <label class="film-details__emoji-label" for="emoji-angry">
+        <img src="./images/emoji/angry.png" width="30" height="30" alt="emoji">
+      </label>
+    </div>
           </ul>
           </div>
         </section>
@@ -101,6 +157,7 @@ export default class FilmDetail extends AbstractSmartComponent {
     this._count = count;
     this._details = details;
     this._closeButtonHandler = null;
+    this._emoji = null;
 
     this._subscribeOnEvents();
 
@@ -111,16 +168,12 @@ export default class FilmDetail extends AbstractSmartComponent {
   }
 
   recoveryListeners() {
-    this.setWatchListButtonClickHandler(this._watchListButtonClickHandler);
+    this.onCloseButtonHandler(this._closeButtonHandler);
     this._subscribeOnEvents();
   }
 
-  rerender() {
-    super.rerender();
-  }
-
   getTemplate() {
-    return filmDetailTemplate(this._film, this._count, this._details);
+    return filmDetailTemplate(this._film, this._count, this._details, this._emoji);
   }
 
   onCloseButtonHandler(handler) {
@@ -143,6 +196,13 @@ export default class FilmDetail extends AbstractSmartComponent {
 
     element.querySelector(`#favorite`).addEventListener(`click`, () => {
       this._isFavorite = !this._isFavorite;
+    });
+
+    element.querySelectorAll(`.film-details__emoji-label`).forEach((it) => {
+      it.addEventListener(`click`, (event) => {
+        this._emoji = event.target.src;
+        this.rerender();
+      });
     });
   }
 }
